moved item logic to its own component
